feat: create material description for gizmo shaders and update MeshLine and Shape materials
