+++ src/extras/gizmo/mesh-line.js
@@ -2,11 +2,11 @@ import { math } from '../../core/math/math.js';
 import { Vec3 } from '../../core/math/vec3.js';
 import { Entity } from '../../framework/entity.js';
 import { BlendState } from '../../platform/graphics/blend-state.js';
-import { SEMANTIC_POSITION } from '../../platform/graphics/constants.js';
 import { CylinderGeometry } from '../../scene/geometry/cylinder-geometry.js';
 import { ShaderMaterial } from '../../scene/materials/shader-material.js';
 import { MeshInstance } from '../../scene/mesh-instance.js';
 import { Mesh } from '../../scene/mesh.js';
+import { materialDesc } from './shaders.js';
 
 /**
  * @import { Color } from '../../core/math/color.js'
@@ -15,62 +15,6 @@ import { Mesh } from '../../scene/mesh.js';
  */
 
 const tmpV1 = new Vec3();
-
-const shaderDesc = {
-    uniqueName: 'mesh-line',
-    attributes: {
-        vertex_position: SEMANTIC_POSITION
-    },
-    vertexGLSL: /* glsl */`
-        attribute vec3 vertex_position;
-    
-        uniform mat4 matrix_model;
-        uniform mat4 matrix_viewProjection;
-    
-        void main(void) {
-            gl_Position = matrix_viewProjection * matrix_model * vec4(vertex_position, 1.0);
-            gl_Position.z = clamp(gl_Position.z, -abs(gl_Position.w), abs(gl_Position.w));
-        }
-    `,
-    fragmentGLSL: /* glsl */`
-        #include "gammaPS"
-    
-        precision highp float;
-    
-        uniform vec4 uColor;
-
-        void main(void) {
-            gl_FragColor = vec4(gammaCorrectOutput(decodeGamma(uColor)), uColor.w);
-        }
-    `,
-    vertexWGSL: /* wgsl */`
-        attribute vertex_position: vec3f;
-
-        uniform matrix_model: mat4x4f;
-        uniform matrix_viewProjection: mat4x4f;
-
-        @vertex
-        fn vertexMain(input: VertexInput) -> VertexOutput {
-            var output: VertexOutput;
-            let pos = vec4f(input.vertex_position, 1.0);
-            output.position = uniform.matrix_viewProjection * uniform.matrix_model * pos;
-            output.position.z = clamp(output.position.z, -abs(output.position.w), abs(output.position.w));
-            return output;
-        }
-    `,
-    fragmentWGSL: /* wgsl */`
-        #include "gammaPS"
-
-        uniform uColor: vec4f;
-
-        @fragment
-        fn fragmentMain(input: FragmentInput) -> FragmentOutput {
-            var output: FragmentOutput;
-            output.color = vec4f(gammaCorrectOutput(decodeGamma(uniform.uColor)), uniform.uColor.w);
-            return output;
-        }
-    `
-};
 
 /**
  * @ignore
@@ -86,7 +30,7 @@ class MeshLine {
      * @type {ShaderMaterial}
      * @private
      */
-    _material = new ShaderMaterial(shaderDesc);
+    _material = new ShaderMaterial(materialDesc);
 
     /**
      * @type {Entity}
