+++ src/extras/gizmo/shape/shape.js
@@ -3,7 +3,7 @@ import { Vec3 } from '../../../core/math/vec3.js';
 import { ShaderMaterial } from '../../../scene/materials/shader-material.js';
 import { MeshInstance } from '../../../scene/mesh-instance.js';
 import { Entity } from '../../../framework/entity.js';
-import { CULLFACE_BACK, SEMANTIC_POSITION } from '../../../platform/graphics/constants.js';
+import { CULLFACE_BACK } from '../../../platform/graphics/constants.js';
 import { BLEND_NORMAL } from '../../../scene/constants.js';
 
 import { COLOR_GRAY } from '../color.js';
@@ -15,6 +15,7 @@ import { ConeGeometry } from '../../../scene/geometry/cone-geometry.js';
 import { PlaneGeometry } from '../../../scene/geometry/plane-geometry.js';
 import { SphereGeometry } from '../../../scene/geometry/sphere-geometry.js';
 import { TorusGeometry } from '../../../scene/geometry/torus-geometry.js';
+import { materialDesc } from '../shaders.js';
 
 /**
  * @import { GraphicsDevice } from '../../../platform/graphics/graphics-device.js';
@@ -31,62 +32,6 @@ const GEOMETRIES = {
     plane: PlaneGeometry,
     sphere: SphereGeometry,
     torus: TorusGeometry
-};
-
-const shaderDesc = {
-    uniqueName: 'axis-shape',
-    attributes: {
-        vertex_position: SEMANTIC_POSITION
-    },
-    vertexGLSL: /* glsl */`
-        attribute vec3 vertex_position;
-    
-        uniform mat4 matrix_model;
-        uniform mat4 matrix_viewProjection;
-    
-        void main(void) {
-            gl_Position = matrix_viewProjection * matrix_model * vec4(vertex_position, 1.0);
-            gl_Position.z = clamp(gl_Position.z, -abs(gl_Position.w), abs(gl_Position.w));
-        }
-    `,
-    fragmentGLSL: /* glsl */`
-        #include "gammaPS"
-    
-        precision highp float;
-    
-        uniform vec4 uColor;
-
-        void main(void) {
-            gl_FragColor = vec4(gammaCorrectOutput(decodeGamma(uColor)), uColor.w);
-        }
-    `,
-    vertexWGSL: /* wgsl */`
-        attribute vertex_position: vec3f;
-
-        uniform matrix_model: mat4x4f;
-        uniform matrix_viewProjection: mat4x4f;
-
-        @vertex
-        fn vertexMain(input: VertexInput) -> VertexOutput {
-            var output: VertexOutput;
-            let pos = vec4f(input.vertex_position, 1.0);
-            output.position = uniform.matrix_viewProjection * uniform.matrix_model * pos;
-            output.position.z = clamp(output.position.z, -abs(output.position.w), abs(output.position.w));
-            return output;
-        }
-    `,
-    fragmentWGSL: /* wgsl */`
-        #include "gammaPS"
-
-        uniform uColor: vec4f;
-
-        @fragment
-        fn fragmentMain(input: FragmentInput) -> FragmentOutput {
-            var output: FragmentOutput;
-            output.color = vec4f(gammaCorrectOutput(decodeGamma(uniform.uColor)), uniform.uColor.w);
-            return output;
-        }
-    `
 };
 
 const tmpG = new Geometry();
@@ -149,7 +94,7 @@ class Shape {
      * @type {ShaderMaterial}
      * @protected
      */
-    _material = new ShaderMaterial(shaderDesc);
+    _material = new ShaderMaterial(materialDesc);
 
     /**
      * The internal disabled state of the shape.
